Updating circle.yml to use jobs and workflows (#48)
The build seems to be failing on circleci due to invalid config. I'm guessing CircleCI's v2, that introduced workflows, deprecates the `stages` config. I downloaded the circleci CLI and validated the config and ran the build locally. Everything seemed to work, except that locally the `restore_cache` and `save_cache` commands failed. I assume this is a quirk of running locally, and the build still worked.